requirements.txt
sqlglot
duckdb
snowflake-connector-python[pandas]
python-dotenv
pyarrow
pandas
streamlit